Moves files to appropriate package
