+++ app/src/main/java/angelosz/catbattlegame/ui/armory/cats_view/DetailedArmoryCatData.kt
@@ -1,4 +1,4 @@
-package angelosz.catbattlegame.ui.armory
+package angelosz.catbattlegame.ui.armory.cats_view
 
 import androidx.annotation.DrawableRes
 import angelosz.catbattlegame.R
